Revert "refactor: extract RevokeDelegationPanel into focused components"
This reverts commit ee611147b3927d3c40511a8ea2546b2216bc0199.
